+++ src/main/java/com/mobitel/data_management/entity/Amc.java
@@ -20,8 +20,8 @@ public class Amc {
     private String contractName;
     private String existingPartner;
     private String userDivision;
-    private double initialCostUSD;
-    private double initialCostLKR;
+    private long initialCostUSD;
+    private long initialCostLKR;
 
     @DateTimeFormat(pattern = "yyyy-MM-dd")
     private Date startDate;
@@ -29,9 +29,9 @@ public class Amc {
     @DateTimeFormat(pattern = "yyyy-MM-dd")
     private Date endDate;
 
-    private double amcValueUSD;
-    private double amcValueLKR;
-    private double amcPercentageUponPurchasePrice;
+    private long amcValueUSD;
+    private long amcValueLKR;
+    private long amcPercentageUponPurchasePrice;
     private String category;
 
     @ManyToOne
